test(lib): Move the 429 handler test to cl.lib.tests
Per review. The behavior under test is the django-ratelimit middleware
plumbing, not a simple_pages page.
